# public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Maw3idy API Documentation'
  description: "Read a business's services and availability, and create or cancel bookings on its behalf."
  version: 1.0.0
servers:
  -
    url: 'http://demo.maw3idy.localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: "Exchange a staff account's credentials for a token with <code>POST /api/v1/tokens</code>. A token belongs to one business: it only works on that business's own subdomain, and only for the abilities it was issued with."
security:
  -
    default: []
paths:
...
  /api/v1/bookings:
    get:
      summary: 'List bookings'
      operationId: listBookings
      description: 'Needs the `bookings:read` ability. Paginated, oldest appointment first.'
      parameters:
        -
          in: query
          name: from
          description: 'Only appointments starting at or after this instant. Must be a valid date.'
          required: false
          schema:
            type: string
            description: 'Only appointments starting at or after this instant. Must be a valid date.'
            examples:
              - '2026-10-05T00:00:00Z'
        -
          in: query
          name: to
          description: 'Only appointments starting at or before this instant. Must be a valid date. Must be a date after or equal to <code>from</code>.'
          required: false
          schema:
            type: string
            description: 'Only appointments starting at or before this instant. Must be a valid date. Must be a date after or equal to <code>from</code>.'
            examples:
              - '2026-10-12T00:00:00Z'
        -
          in: query
          name: staff_id
          description: "Only this staff member's appointments."
          required: false
          schema:
            type: integer
            description: "Only this staff member's appointments."
            examples:
              - 1
        -
          in: query
          name: status
          description: 'Either `confirmed` or `cancelled`.'
          required: false
          schema:
            type: string
            description: 'Either `confirmed` or `cancelled`.'
            enum:
              - confirmed
              - cancelled
            examples:
              - confirmed
        -
          in: query
          name: per_page
          description: 'Results per page, 1 to 100. Must be at least 1. Must not be greater than 100.'
          required: false
          schema:
            type: integer
            description: 'Results per page, 1 to 100. Must be at least 1. Must not be greater than 100.'
            examples:
              - 25
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        reference:
                          type: string
                          examples:
                            - MW-7K3P9Q
                        status:
                          type: string
                          examples:
                            - confirmed
                        starts_at:
                          type: string
                          examples:
                            - '2026-10-05T07:00:00+00:00'
                        ends_at:
                          type: string
                          examples:
                            - '2026-10-05T07:30:00+00:00'
                        duration_minutes:
                          type: integer
                          examples:
                            - 30
                        buffer_after_minutes:
                          type: integer
                          examples:
                            - 5
                        price:
                          type: string
                          examples:
                            - '80.00'
                        currency:
                          type: string
                          examples:
                            - SAR
                        customer:
                          type: object
                          properties:
                            name:
                              type: string
                              examples:
                                - 'Layla A.'
                            phone:
                              type: string
                              examples:
                                - '+966501112222'
                            email:
                              type:
                                - string
                                - 'null'
                              examples:
                                - null
                        cancelled_at:
                          type:
                            - string
                            - 'null'
                          examples:
                            - null
                        cancellation_reason:
                          type:
                            - string
                            - 'null'
                          examples:
                            - null
                        created_at:
                          type: string
                          examples:
                            - '2026-09-28T09:14:00+00:00'
                    examples:
                      -
                        -
                          reference: MW-7K3P9Q
                          status: confirmed
                          starts_at: '2026-10-05T07:00:00+00:00'
                          ends_at: '2026-10-05T07:30:00+00:00'
                          duration_minutes: 30
                          buffer_after_minutes: 5
                          price: '80.00'
                          currency: SAR
                          customer:
                            name: 'Layla A.'
                            phone: '+966501112222'
                            email: null
                          cancelled_at: null
                          cancellation_reason: null
                          created_at: '2026-09-28T09:14:00+00:00'
                  meta:
                    type: object
                    properties:
                      timezone:
                        type: string
                        examples:
                          - Asia/Riyadh
                      current_page:
                        type: integer
                        examples:
                          - 1
                      per_page:
                        type: integer
                        examples:
                          - 25
                      total:
                        type: integer
                        examples:
                          - 1
                examples:
                  -
                    data:
                      -
                        reference: MW-7K3P9Q
                        status: confirmed
                        starts_at: '2026-10-05T07:00:00+00:00'
                        ends_at: '2026-10-05T07:30:00+00:00'
                        duration_minutes: 30
                        buffer_after_minutes: 5
                        price: '80.00'
                        currency: SAR
                        customer:
                          name: 'Layla A.'
                          phone: '+966501112222'
                          email: null
                        cancelled_at: null
                        cancellation_reason: null
                        created_at: '2026-09-28T09:14:00+00:00'
                    meta:
                      timezone: Asia/Riyadh
                      current_page: 1
                      per_page: 25
                      total: 1
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - 'This token is not allowed to do that.'
                examples:
                  -
                    message: 'This token is not allowed to do that.'
      tags:
        - Bookings
    post:
      summary: 'Create a booking'
      operationId: createABooking
      description: "Needs the `bookings:write` ability.\n\nOmit `staff_id` to let the platform pick anyone who offers the service\nand is free. The slot is verified again under a lock before the booking\nis written, so two callers racing for the same time cannot both win: the\nloser gets `409`.\n\nThe service's duration, buffer and price are copied onto the booking, so\nlater changes to the service never rewrite an existing appointment."
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      reference:
                        type: string
                        examples:
                          - MW-7K3P9Q
                      status:
                        type: string
                        examples:
                          - confirmed
                      starts_at:
                        type: string
                        examples:
                          - '2026-10-05T07:00:00+00:00'
                      ends_at:
                        type: string
                        examples:
                          - '2026-10-05T07:30:00+00:00'
                      duration_minutes:
                        type: integer
                        examples:
                          - 30
                      buffer_after_minutes:
                        type: integer
                        examples:
                          - 5
                      price:
                        type: string
                        examples:
                          - '80.00'
                      currency:
                        type: string
                        examples:
                          - SAR
                      customer:
                        type: object
                        properties:
                          name:
                            type: string
                            examples:
                              - 'Layla A.'
                          phone:
                            type: string
                            examples:
                              - '+966501112222'
                          email:
                            type:
                              - string
                              - 'null'
                            examples:
                              - null
                      cancelled_at:
                        type:
                          - string
                          - 'null'
                        examples:
                          - null
                      cancellation_reason:
                        type:
                          - string
                          - 'null'
                        examples:
                          - null
                      created_at:
                        type: string
                        examples:
                          - '2026-09-28T09:14:00+00:00'
                  meta:
                    type: object
                    properties:
                      timezone:
                        type: string
                        examples:
                          - Asia/Riyadh
                examples:
                  -
                    data:
                      reference: MW-7K3P9Q
                      status: confirmed
                      starts_at: '2026-10-05T07:00:00+00:00'
                      ends_at: '2026-10-05T07:30:00+00:00'
                      duration_minutes: 30
                      buffer_after_minutes: 5
                      price: '80.00'
                      currency: SAR
                      customer:
                        name: 'Layla A.'
                        phone: '+966501112222'
                        email: null
                      cancelled_at: null
                      cancellation_reason: null
                      created_at: '2026-09-28T09:14:00+00:00'
                    meta:
                      timezone: Asia/Riyadh
        409:
          description: 'The time was taken while the caller was deciding.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - 'The slot starting at 2026-10-05T07:00:00+00:00 is no longer available.'
                examples:
                  -
                    message: 'The slot starting at 2026-10-05T07:00:00+00:00 is no longer available.'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - 'The customer name field is required.'
                  errors:
                    type: object
                    properties:
                      customer_name:
                        type: array
                        items:
                          type: string
                        examples:
                          -
                            - 'The customer name field is required.'
                examples:
                  -
                    message: 'The customer name field is required.'
                    errors:
                      customer_name:
                        - 'The customer name field is required.'
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service_id:
                  type: integer
                  description: 'A bookable service at this business. Must match an existing stored value.'
                  examples:
                    - 1
                staff_id:
                  type:
                    - integer
                    - 'null'
                  description: 'Who should take it. Omit for anyone available. Must match an existing stored value.'
                  examples:
                    - 1
                starts_at:
                  type: string
                  description: 'When the appointment starts. Any offset; stored as UTC. Must be a valid date.'
                  examples:
                    - '2026-10-05T07:00:00+00:00'
                customer_name:
                  type: string
                  description: "The customer's name. Must not be greater than 100 characters."
                  examples:
                    - 'Layla A.'
                customer_phone:
                  type: string
                  description: '8 to 15 digits, with or without a country code. Must match the regex /^\+?[0-9]{8,15}$/.'
                  examples:
                    - '+966501112222'
                customer_email:
                  type:
                    - string
                    - 'null'
                  description: 'Optional. Without it the customer cannot be emailed a confirmation. Must be a valid email address. Must not be greater than 255 characters.'
                  examples:
                    - layla@example.com
              required:
                - service_id
                - starts_at
                - customer_name
                - customer_phone
    parameters:
      -
        in: path
        name: tenant
        description: "The business's subdomain."
        required: true
        schema:
          type: string
          examples:
            - demo
...